Move the header in the layout
