+++ app/about/page.tsx
@@ -1,11 +1,10 @@
 import Button from "@mui/material/Button";
-import { Layout } from "../components/Layout";
 
 export default function About() {
   return (
-    <Layout>
+    <div>
       Hello About
       <Button>Test</Button>
-    </Layout>
+    </div>
   );
 }
