+++ app/dossiers/page.tsx
@@ -1,12 +1,6 @@
-import { Layout } from "../components/Layout";
-
 import Link from "next/link";
 
 export default function Dossiers() {
-  return (
-    <Layout>
-      {/* TODO: match old website url ":id/dossiers/:id" */}
-      <Link href="dossiers/dossier">Lien vers un dossier</Link>
-    </Layout>
-  );
+  // TODO: match old website url ":id/dossiers/:id"
+  return <Link href="dossiers/dossier">Lien vers un dossier</Link>;
 }
